+++ src/components/pie-semi-circle-chart/index.ts
@@ -1,0 +1,2 @@
+export { default as PieSemiCircleChart } from './pie-semi-circle-chart';
+export type { PieSemiCircleChartProps, ArcData } from './pie-semi-circle-chart';